chore!: explicitly set template entry
Checking if a template entry exists (introduced in e5c170d3762e984129c818eff8be7ccaed3a9313)
wasn't working properly (see #12). This could be fixed on the php side
by checking if we're in dev mode, but could still raise an error in the
console during dev, because we don't know if a template entry exists or
not during development.

Migration: in each template php file, explicitly pass the entry name to
the snippet: `snippet('header', ['entry' => 'home'])`. If no entry is
passed, only the shared entry is loaded.